+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for differential GPS last time
 - Rework library CMake with removed INTERFACE type
+- Add `time_valid` and `date_valid` to give information if information from modem is valid
 
 ## v2.2.0